+++ packages/plugin-css/src/utils/format-font-names.ts
@@ -1,0 +1,4 @@
+/** format font stack */
+export function formatFontNames(fontNames: string[]): string {
+  return fontNames.map((n) => (n.includes(' ') ? `"${n}"` : n)).join(', ');
+}
