+++ packages/plugin-css/src/utils/is-token-match.ts
@@ -1,0 +1,13 @@
+import wcmatch from '../wildcard-match.js';
+
+/** match token against globs */
+export function isTokenMatch(tokenID: string, globPatterns: string[]): boolean {
+  let isMatch = false;
+  for (const tokenMatch of globPatterns) {
+    if (wcmatch(tokenMatch)(tokenID)) {
+      isMatch = true;
+      break;
+    }
+  }
+  return isMatch;
+}
