+++ packages/plugin-css/tsconfig.json
@@ -3,5 +3,6 @@
   "compilerOptions": {
     "outDir": "./dist/"
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["src/**/*.test.ts"]
 }
